attestation: add snp package
The package holds code shared between SNP-based
attestation implementations on AWS and Azure .
