+++ internal/attestation/azure/snp/BUILD.bazel
@@ -15,11 +15,11 @@ go_library(
     deps = [
         "//internal/attestation",
         "//internal/attestation/idkeydigest",
+        "//internal/attestation/snp",
         "//internal/attestation/variant",
         "//internal/attestation/vtpm",
         "//internal/cloud/azure",
         "//internal/config",
-        "//internal/constants",
         "@com_github_edgelesssys_go_azguestattestation//maa",
         "@com_github_google_go_sev_guest//abi",
         "@com_github_google_go_sev_guest//kds",
@@ -48,9 +48,10 @@ go_test(
     }),
     deps = [
         "//internal/attestation",
-        "//internal/attestation/azure/snp/testdata",
         "//internal/attestation/idkeydigest",
         "//internal/attestation/simulator",
+        "//internal/attestation/snp",
+        "//internal/attestation/snp/testdata",
         "//internal/attestation/vtpm",
         "//internal/config",
         "//internal/logger",
