+++ internal/attestation/snp/testdata/attestation.bin
