+++ internal/attestation/snp/testdata/runtimedata.bin
