+++ internal/attestation/snp/testdata/testdata.go
