+++ internal/attestation/snp/testdata/vcek.cert
