+++ internal/attestation/snp/testdata/vcek.pem
